% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.evoTSfit.OUBM.R
\name{as.evoTSfit.OUBM}
\alias{as.evoTSfit.OUBM}
\title{Class for fit to evolutionary sequence (time-series) models}
\usage{
as.evoTSfit.OUBM(logL, parameters, modelName, method, K, n, se)
}
\arguments{
\item{logL}{log-likelihood of model}

\item{parameters}{maximum-likelihood estimates of the parameters}

\item{modelName}{description of the model}

\item{method}{the parameterization used: Joint}

\item{K}{number of parameters in the model}

\item{n}{sample size}

\item{se}{standard errors of parameter estimates}
}
\description{
A function that combines useful information summarizing model fit.
}
\details{
This function is used by the model-fitting routines for the univariate Ornstein-Uhlenbeck model where the optimum evolves as an Unbiased Random Walk to create standardized output
}
\note{
This function is not likely to be called directly by the user.
}
\author{
Kjetil Lysne Voje
}
