% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logL.joint.accel.decel.single.R.zero.corr.R
\name{logL.joint.accel.decel.single.R.zero.corr}
\alias{logL.joint.accel.decel.single.R.zero.corr}
\title{Log-likelihoods for evolutionary models}
\usage{
logL.joint.accel.decel.single.R.zero.corr(
  init.par,
  y,
  m,
  n,
  anc.values,
  ta,
  se_vec
)
}
\arguments{
\item{init.par}{initial (starting) parameters values}

\item{y}{vector containing all trait values from all traits}

\item{m}{number of traits}

\item{n}{number of populations}

\item{anc.values}{initial values for the ancestral trait values}

\item{ta}{matrix of minimum times for each pair of time points}

\item{se_vec}{pre-computed sampling error vector}
}
\value{
The log-likelihood of the parameter estimates, given the data.
}
\description{
Returns log-likelihood for a multivariate Unbiased Random Walk model with
uncorrelated changes and with accelerating or decelerating rates of evolution through time.
}
\details{
In general, users will not be access these functions directly, but instead use the
optimization functions, which use these functions to find the best-supported parameter values.
}
\author{
Kjetil Lysne Voje
}
