% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logL.joint.accel_decel.R
\name{logL.joint.accel_decel}
\alias{logL.joint.accel_decel}
\title{Log-likelihoods for evolutionary models}
\usage{
logL.joint.accel_decel(p, y)
}
\arguments{
\item{p}{parameters of the model to be optimized}

\item{y}{a paleoTS object}
}
\value{
The log-likelihood of the parameter estimates, given the data.
}
\description{
Returns log-likelihood for an Unbiased Random Walk with an accelerating or decelerating rate of change through time.
}
\details{
In general, users will not be access these functions directly, but instead use the optimization functions, which use these functions to find the best-supported parameter values.
}
\author{
Kjetil Lysne Voje
}
