% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logL.joint.Stasis.OU.R
\name{logL.joint.Stasis.OU}
\alias{logL.joint.Stasis.OU}
\title{Log-likelihoods for evolutionary models}
\usage{
logL.joint.Stasis.OU(p, y, gg)
}
\arguments{
\item{p}{parameters of the model to be optimized}

\item{y}{a paleoTS object}

\item{gg}{numeric vector indicating membership of each sample in a segment}
}
\value{
The log-likelihood of the parameter estimates, given the data.
}
\description{
Returns log-likelihood for a model with stasis in the first segment and an Ornstein-Uhlenbeck process in the second segment. .
}
\details{
In general, users will not be access these functions directly, but instead use the optimization functions, which use these functions to find the best-supported parameter values.
}
\author{
Kjetil Lysne Voje
}
