% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.multivar.evoTS.R
\name{make.multivar.evoTS}
\alias{make.multivar.evoTS}
\title{Makes a multivariate data set of}
\usage{
make.multivar.evoTS(...)
}
\arguments{
\item{...}{two or more univariate evolutionary sequences (time-series) in the format used by paleoTS, passed as individual arguments.}
}
\value{
a multivariate evoTS object that can be analysed with functions fitting multivariate models (e.g. fit.multivariate.OU, fit.multivariate.URW)
}
\description{
Function to make a multivariate data set consisting of two or more evolutionary sequences (time-series).
}
\details{
See the function as.paleoTS for details. See also read.paleoTS, which is often a more convenient way for getting the relevant data from text files.
}
\examples{
## Generate two evolutionary sequences (time-series)
x1 <- paleoTS::sim.GRW(60)
x2 <- paleoTS::sim.GRW(60)


## Make a multivariate data set
x1_x2<-make.multivar.evoTS(x1, x2)

}
\author{
Kjetil Lysne Voje
}
