% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.and.logl.pairs.R
\name{opt.joint.URW.Stasis}
\alias{opt.joint.URW.Stasis}
\title{Optimization and log-likelihoods for pairs of models.}
\usage{
opt.joint.URW.Stasis(
  y,
  gg,
  cl = list(fnscale = -1),
  pool = TRUE,
  meth = "L-BFGS-B",
  hess = FALSE
)
}
\arguments{
\item{y}{a paleoTS object.}

\item{gg}{numeric vector indicating membership of each sample in segments}

\item{cl}{control list to be passed to optim}

\item{pool}{logical indicating whether to pool variances across samples}

\item{meth}{optimization method, passed to function optim. Default is "L-BFGS-B".}

\item{hess}{logical, indicating whether to calculate standard errors from the Hessian matrix.}
}
\description{
A collections of functions that serves the function fit.mode.shift. See fit.mode.shift for info.
}
\details{
In general, users will not be access these functions directly, but instead use the wrapper function, which use these functions to find the best-supported parameter values.
}
\note{
This function is not likely to be called directly by the user.
}
\author{
Kjetil Lysne Voje
}
