% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logl.joint.multi.OUOU.R
\name{logL.joint.multi.OUOU}
\alias{logL.joint.multi.OUOU}
\title{Log-likelihoods for evolutionary models}
\usage{
logL.joint.multi.OUOU(init.par, yy, A.matrix, R.matrix)
}
\arguments{
\item{init.par}{initial (starting) parameters values}

\item{yy}{a multivariate evoTS object}

\item{A.matrix}{the pull matrix.}

\item{R.matrix}{the drift matrix..}
}
\value{
The log-likelihood of the parameter estimates, given the data.
}
\description{
Returns log-likelihood for a multivariate Ornstein-Uhlenbeck model.
}
\details{
In general, users will not be access these functions directly, but instead use the optimization functions, which use these functions to find the best-supported parameter values.
}
\author{
Kjetil Lysne Voje
}
