% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.multivar.evoTS.R
\name{make.multivar.evoTS}
\alias{make.multivar.evoTS}
\title{Makes a multivariate data set of}
\usage{
make.multivar.evoTS(
  evoTS.1 = NULL,
  evoTS.2 = NULL,
  evoTS.3 = NULL,
  evoTS.4 = NULL,
  evoTS.5 = NULL,
  evoTS.6 = NULL,
  evoTS.7 = NULL,
  evoTS.8 = NULL,
  evoTS.9 = NULL,
  evoTS.10 = NULL
)
}
\arguments{
\item{evoTS.1}{an univariate evolutionary sequences (time-series) on the format used in paleoTS}

\item{evoTS.2}{an univariate evolutionary sequences (time-series) on the format used in paleoTS}

\item{evoTS.3}{an univariate evolutionary sequences (time-series) on the format used in paleoTS (optional)}

\item{evoTS.4}{an univariate evolutionary sequences (time-series) on the format used in paleoTS (optional)}

\item{evoTS.5}{an univariate evolutionary sequences (time-series) on the format used in paleoTS (optional)}

\item{evoTS.6}{an univariate evolutionary sequences (time-series) on the format used in paleoTS (optional)}

\item{evoTS.7}{an univariate evolutionary sequences (time-series) on the format used in paleoTS (optional)}

\item{evoTS.8}{an univariate evolutionary sequences (time-series) on the format used in paleoTS (optional)}

\item{evoTS.9}{an univariate evolutionary sequences (time-series) on the format used in paleoTS (optional)}

\item{evoTS.10}{an univariate evolutionary sequences (time-series) on the format used in paleoTS (optional)}
}
\value{
a multivariate evoTS object that can be analysed with functions fitting multivariate models (e.g. fit.multivariate.OU, fit.multivariate.URW)
}
\description{
Function to make a multivariate data set consisting of two or more evolutionary sequences (time-series).
}
\details{
See the function as.paleoTS for details. See also read.paleoTS, which is often a more convenient way for getting the relevant data from text files.
}
\examples{
## Generate two evolutionary sequences (time-series)
x1 <- paleoTS::sim.GRW(60)
x2 <- paleoTS::sim.GRW(60)


## Make a multivariate data set
x1_x2<-make.multivar.evoTS(x1, x2)

}
\author{
Kjetil Lysne Voje
}
